Rebase Selenium Buck on Facebook Buck
This pulls up to the most recent source of Facebook's
Buck, and retrofits as much of our stuff as we feel is
necessary.

Note that we pull in a couple of PRs that have not yet
been accepted to Facebook's Buck in order to get things
like source jar and docs publishing via maven.